Add files via upload
Hi! I had issues with uploading the frontend file since it's such a large file. I could not make a zip out of it since it had characters that were not allowed. So I uploaded all except for the node_modules (I did not edit anything in this file.  

Also, my code is not working 100%: the front end gave me errors and at the end of the code it's not the correct code (it was my first time writing api's). I would love to know or have some insights on why it's not working and what I should do differently! Thanks